I have expanded upon the work James House did to generate stack traces for abandoned db connections. (Thanks James)
The patch he provided was refactored and some new features added.

Three new DBCP parameters were added:

removeAbandoned - True or false. If true abandoned db connections
are removed after the removeAbandonedTimout is exceeded if the dbcp
is nearing exhaustion.

removeAbandonedTimeout - Time in seconds since a connection was last
used before it is considered abandoned.

logAbandoned - True or false.  If true Exception stack traces are
created so that the source of an abandoned db connection can be logged.

While reviewing the code I noticed that Statements and ResultSets were
not being closed when a Connection was closed (recycled).  This behaviour
differs from the JDBC spec which states that the underlying Statements
and ResultSets should be closed when a Connection is closed.  This
patch tracks Statements and ResultSets so that when the Connection
which created them is closed they are closed also.

This patch should be backward compatible with any code which uses
the existing DBCP.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/dbcp/trunk@131884 13f79535-47bb-0310-9956-ffa450edef68