Fix DelegatingConnection readOnly and autoCommit caching mechanism
Previously, we would cache the input of setReadOnly and setAutoCommit directly without checking if the underlying connection succeeded in setting the configuration parameter. This would result in wrong results from isReadOnly and getAutoCommit if the underlying connection didn't support read-only/non-read-only or auto-commit/non-auto-commit connections, as with SqlServerConnection.

This fixes the issue by caching what is return by isReadOnly and getAutoCommit from the underlying connection.